refactor(es): one openSearchStream + shared applyCollapse for the three search entry points
searchRaw was a ~50-line verbatim copy of searchStream and both re-copied
search()'s collapse guard — the mirror invariant is now enforced by code.
